AOC2021 - Day #5

+++ src/main/java/com/milossamek/aoc2021/solutions/day2/SubmarineCommandCalculator.java
@@ -23,7 +23,7 @@ public class SubmarineCommandCalculator extends ResourceLoader<SubmarineCommand>
                 case UP -> depth.addAndGet(-command.getMove());
             }
         });
-        return horizontal.get() * depth.get();
+        return (long) horizontal.get() * depth.get();
     }
 
     @Override
